third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-hidden-iframe-zero-size-expected.txt

CONSOLE WARNING: Resource requests whose URLs contain raw newline characters are deprecated, and may be blocked in M60, around August 2017. Please remove newlines from places like element attribute values in order to continue loading those resources. See https://www.chromestatus.com/features/5735596811091968 for more details.
This is link_1.


This is 
an iframe with size 0.
This is link_2.

PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "end"
This test is to test that an iframe with size 0 does not get focus, and neither do its descendants.
